Kill dependency error
